Use singleton package manager for global apis
