+++ micropip/package_index.py
@@ -16,7 +16,6 @@ from .externals.mousebender.simple import from_project_details_html
 from .wheelinfo import WheelInfo
 
 DEFAULT_INDEX_URLS = ["https://pypi.org/simple"]
-INDEX_URLS = DEFAULT_INDEX_URLS
 
 _formatter = string.Formatter()
 
@@ -232,8 +231,8 @@ def _select_parser(content_type: str, pkgname: str) -> Callable[[str], ProjectIn
 
 async def query_package(
     name: str,
+    index_urls: list[str] | str,
     fetch_kwargs: dict[str, Any] | None = None,
-    index_urls: list[str] | str | None = None,
 ) -> ProjectInfo:
     """
     Query for a package from package indexes.
@@ -242,16 +241,13 @@ async def query_package(
     ----------
     name
         Name of the package to search for.
-    fetch_kwargs
-        Keyword arguments to pass to the fetch function.
     index_urls
         A list of URLs or a single URL to use as the package index.
-        If None, the default index URL is used.
-
         If a list of URLs is provided, it will be tried in order until
         it finds a package. If no package is found, an error will be raised.
+    fetch_kwargs
+        Keyword arguments to pass to the fetch function.
     """
-    global INDEX_URLS
 
     _fetch_kwargs = fetch_kwargs.copy() if fetch_kwargs else {}
 
@@ -263,9 +259,7 @@ async def query_package(
         "accept", "application/vnd.pypi.simple.v1+json, */*;q=0.01"
     )
 
-    if index_urls is None:
-        index_urls = INDEX_URLS
-    elif isinstance(index_urls, str):
+    if isinstance(index_urls, str):
         index_urls = [index_urls]
 
     for url in index_urls:
