+++ tests/test_package_index.py
@@ -1,7 +1,6 @@
 import pytest
 from conftest import TEST_PYPI_RESPONSE_DIR, _read_gzipped_testfile
 
-import micropip._commands.index_urls as index_urls
 import micropip.package_index as package_index
 
 
@@ -87,24 +86,6 @@ def test_project_info_equal(name):
             assert f_json.url == f_simple_json.url
             assert f_json.version == f_simple_json.version
             assert f_json.sha256 == f_simple_json.sha256
-
-
-def test_set_index_urls():
-    default_index_urls = package_index.DEFAULT_INDEX_URLS
-    assert package_index.INDEX_URLS == default_index_urls
-
-    valid_url1 = "https://pkg-index.com/{package_name}/json/"
-    valid_url2 = "https://another-pkg-index.com/{package_name}"
-    valid_url3 = "https://another-pkg-index.com/simple/"
-    try:
-        index_urls.set_index_urls(valid_url1)
-        assert package_index.INDEX_URLS == [valid_url1]
-
-        index_urls.set_index_urls([valid_url1, valid_url2, valid_url3])
-        assert package_index.INDEX_URLS == [valid_url1, valid_url2, valid_url3]
-    finally:
-        index_urls.set_index_urls(default_index_urls)
-        assert package_index.INDEX_URLS == default_index_urls
 
 
 def test_contain_placeholder():
